Pull `CoordinateFrame` into its own module (`_core`)
